# docs/swagger.yaml
basePath: /
definitions:
...
  auth.ChangePasswordRequest:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  auth.CurrentUser:
    properties:
      campusId:
        type: integer
      campusName:
        type: string
      dataScope:
        type: string
      email:
        type: string
      featureFlags:
        items:
          type: string
        type: array
      id:
        type: integer
      menuPermissions:
        items:
          type: string
        type: array
      mustChangePassword:
        type: boolean
      organizationId:
        type: integer
      organizationName:
        type: string
      permissions:
        items:
          type: string
        type: array
      phone:
        type: string
      realName:
        type: string
      remark:
        type: string
      roleCode:
        type: string
      status:
        type: string
      subscriptionStatus:
        type: string
      username:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginResponse:
    properties:
      expiresIn:
        type: integer
      refreshToken:
        type: string
      session:
        $ref: '#/definitions/auth.SessionInfo'
      token:
        type: string
      tokenType:
        type: string
      userInfo:
        $ref: '#/definitions/auth.CurrentUser'
    type: object
  auth.PasswordRecoveryResetRequest:
    properties:
      challengeId:
        type: string
      newPassword:
        type: string
      secondFactorCode:
        type: string
      verificationCode:
        type: string
    required:
    - challengeId
    - newPassword
    - verificationCode
    type: object
  auth.PasswordRecoveryResetResponse:
    properties:
      recoveredAt:
        type: string
      recoveryChallengeId:
        type: string
      requiresFreshLogin:
        type: boolean
      revokedSessionCount:
        type: integer
      success:
        type: boolean
    type: object
  auth.PasswordRecoveryStartRequest:
    properties:
      channel:
        type: string
      secondFactorChannel:
        type: string
      username:
        type: string
    required:
    - channel
    - username
    type: object
  auth.PasswordRecoveryStartResponse:
    properties:
      accepted:
        type: boolean
      challengeId:
        type: string
      expiresIn:
        type: integer
      message:
        type: string
      primaryChannel:
        type: string
      primaryTargetMasked:
        type: string
      riskFlags:
        items:
          type: string
        type: array
      secondFactorChannel:
        type: string
      secondFactorRequired:
        type: boolean
      secondTargetMasked:
        type: string
    type: object
  auth.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  auth.SessionInfo:
    properties:
      campusId:
        type: integer
      clientIP:
        type: string
      createdAt:
        type: string
      current:
        type: boolean
      deviceId:
        type: string
      deviceName:
        type: string
      id:
        type: string
      lastRefreshedAt:
        type: string
      lastSeenAt:
        type: string
      organizationId:
        type: integer
      revokedAt:
        type: string
      revokedReason:
        type: string
      riskFlags:
        items:
          type: string
        type: array
      roleCode:
        type: string
      status:
        type: string
      suspicious:
        type: boolean
      updatedAt:
        type: string
      userAgent:
        type: string
    type: object
  auth.SessionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/auth.SessionInfo'
        type: array
      total:
        type: integer
    type: object
  code.Failure:
    properties:
      code:
        description: 业务码
        type: integer
      message:
        description: 描述信息
        type: string
    type: object
...
info:
  contact: {}
  title: Edu Schedule System 接口文档
  version: v0.0.1
paths:
...
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: 用于首次登录强制改密，也可用于当前登录用户主动修改密码。
      parameters:
      - description: 请求参数
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/auth.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.CurrentUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - BearerAuth: []
      summary: 修改当前用户密码
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 使用 sys_user 用户表进行正式登录，签发 JWT。
      parameters:
      - description: 请求参数
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      summary: 正式账号密码登录
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: 撤销当前 access token，并在提供 refresh token 时一并使其失效。
      parameters:
      - description: 可选 refresh token
        in: header
        name: X-Refresh-Token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - BearerAuth: []
      summary: 当前用户登出
      tags:
      - Auth
  /api/v1/auth/logout-all:
    post:
      consumes:
      - application/json
      description: 撤销当前用户的全部会话，包含当前 access token 对应会话。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - BearerAuth: []
      summary: 注销全部其他会话
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: 返回当前会话对应的用户信息。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.CurrentUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - BearerAuth: []
      summary: 获取当前登录用户信息
      tags:
      - Auth
  /api/v1/auth/password-recovery/reset:
    post:
      consumes:
      - application/json
      description: 校验恢复挑战后重置密码，并撤销历史会话。
      parameters:
      - description: 请求参数
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordRecoveryResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.PasswordRecoveryResetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      summary: 完成密码恢复并重置密码
      tags:
      - Auth
  /api/v1/auth/password-recovery/start:
    post:
      consumes:
      - application/json
      description: 通过预设恢复渠道发起找回密码挑战。
      parameters:
      - description: 请求参数
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordRecoveryStartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.PasswordRecoveryStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      summary: 发起密码恢复
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh token 刷新 access token，并复用当前会话。
      parameters:
      - description: 请求参数
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      summary: 刷新访问令牌
      tags:
      - Auth
  /api/v1/auth/sessions:
    get:
      consumes:
      - application/json
      description: 返回当前用户的可见登录会话列表。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SessionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - BearerAuth: []
      summary: 列出当前用户会话
      tags:
      - Auth
  /api/v1/auth/sessions/{id}:
    delete:
      consumes:
      - application/json
      description: 撤销当前用户名下的指定会话。
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/code.Failure'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - BearerAuth: []
      summary: 撤销指定会话
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"